Changed API to use a single variable `event`.
event is a namedtuple for plugin the API. Updated plugins to conform.

Codebase better conforms to '\n' and 79 char line width. So much to do.

This commit is a checkpoint. Plugins will be changed to use
self.bot instead of self.protocol. Helpers in bot will added.
This means self.bot.respond instead of self.protocol.respond
or self.bot.protocol.respond, but it also means removing
bot references from the protocol!